Rename syntax object
